# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SESS - Demo Application",
        "description": "This is a sample server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "jccm@sess-corp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9090/DemoProducto"
        }
    ],
    "paths": {
        "/api": {},
        "/api/product": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Listar productos",
                "operationId": "df8cd64ed264db3f7e66ff80a6fc0cfe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Nuevo Producto",
                "operationId": "113cc8ff5ab4cb049f1c4cd9a62c6d79",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Obtener producto",
                "operationId": "a96e41651909cacf69b65bb2f1b620f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Actualizar producto",
                "operationId": "b9409cb8d9b6ba0233a6d4c063961aad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponseData"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Eliminar producto",
                "operationId": "43fd930b777f8935961a113656b3d4a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponseData"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
            "ResponseData": {
                "properties": {
                    "message": {
                        "type": "string",
                        "format": "string"
                    },
                    "entity": {
                        "type": "object",
                        "format": "object"
                    },
                    "status": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}